requirements.txt
langchain==0.3.19
langchain_community==0.3.17
langchain_huggingface==0.1.2
langchain_mistralai==0.2.6
mistralai==1.5.0
pandas==2.2.3
sentence-transformers
faiss-cpu
python-dotenv==1.0.1
